Move estimateSignedTxSize etc to separate module
