+++ lib/wallet/cardano-wallet.cabal
@@ -412,6 +412,7 @@ library
     Cardano.Wallet.Write.Tx.Balance
     Cardano.Wallet.Write.Tx.Gen
     Cardano.Wallet.Write.Tx.Redeemers
+    Cardano.Wallet.Write.Tx.Sign
     Cardano.Wallet.Write.Tx.SizeEstimation
     Cardano.Wallet.Write.Tx.TimeTranslation
     Cardano.Wallet.Write.UTxOAssumptions
